+++ envoy/network/BUILD
@@ -64,6 +64,15 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
+    name = "dns_resolver_interface",
+    hdrs = ["dns_resolver.h"],
+    deps = [
+        "//envoy/api:api_interface",
+        "//source/common/config:utility_lib",
+    ],
+)
+
+envoy_cc_library(
     name = "drain_decision_interface",
     hdrs = ["drain_decision.h"],
     external_deps = ["abseil_base"],
