+++ source/common/event/BUILD
@@ -43,7 +43,6 @@ envoy_cc_library(
         "//envoy/network:listener_interface",
         "//source/common/common:assert_lib",
         "//source/common/common:thread_lib",
-        "//source/common/config:utility_lib",
         "//source/common/filesystem:watcher_lib",
         "//source/common/network:connection_lib",
         "//source/common/network:listener_lib",
