+++ source/common/network/dns_resolver/BUILD
@@ -9,12 +9,11 @@ licenses(["notice"])  # Apache 2
 envoy_package()
 
 envoy_cc_library(
-    name = "dns_factory_lib",
-    srcs = ["dns_factory.cc"],
-    hdrs = ["dns_factory.h"],
+    name = "dns_factory_util_lib",
+    srcs = ["dns_factory_util.cc"],
+    hdrs = ["dns_factory_util.h"],
     deps = [
-        "//envoy/api:api_interface",
-        "//source/common/config:utility_lib",
+        "//envoy/network:dns_resolver_interface",
         "@envoy_api//envoy/config/bootstrap/v3:pkg_cc_proto",
         "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/common/dynamic_forward_proxy/v3:pkg_cc_proto",
