+++ source/extensions/common/wasm/wasm.cc
@@ -6,7 +6,7 @@
 #include "envoy/event/deferred_deletable.h"
 
 #include "source/common/common/logger.h"
-#include "source/common/network/dns_resolver/dns_factory.h"
+#include "source/common/network/dns_resolver/dns_factory_util.h"
 #include "source/extensions/common/wasm/plugin.h"
 #include "source/extensions/common/wasm/stats_handler.h"
