+++ source/extensions/filters/udp/dns_filter/BUILD
@@ -39,6 +39,7 @@ envoy_cc_library(
         "//source/common/config:datasource_lib",
         "//source/common/network:address_lib",
         "//source/common/network:utility_lib",
+        "//source/common/network/dns_resolver:dns_factory_util_lib",
         "//source/common/protobuf:message_validator_lib",
         "//source/common/runtime:runtime_lib",
         "//source/common/upstream:cluster_manager_lib",
