+++ source/extensions/network/dns_resolver/apple/BUILD
@@ -29,7 +29,7 @@ envoy_cc_extension(
         "//source/common/common:linked_object",
         "//source/common/network:address_lib",
         "//source/common/network:utility_lib",
-        "//source/common/network/dns_resolver:dns_factory_lib",
+        "//source/common/network/dns_resolver:dns_factory_util_lib",
         "//source/common/singleton:threadsafe_singleton",
     ],
 )
