+++ source/extensions/network/dns_resolver/apple/apple_dns_impl.cc
@@ -15,7 +15,7 @@
 #include "source/common/common/assert.h"
 #include "source/common/common/fmt.h"
 #include "source/common/network/address_impl.h"
-#include "source/common/network/dns_resolver/dns_factory.h"
+#include "source/common/network/dns_resolver/dns_factory_util.h"
 #include "source/common/network/utility.h"
 
 #include "absl/strings/str_join.h"
