+++ source/server/config_validation/server.h
@@ -16,7 +16,7 @@
 #include "source/common/common/assert.h"
 #include "source/common/common/random_generator.h"
 #include "source/common/grpc/common.h"
-#include "source/common/network/dns_resolver/dns_factory.h"
+#include "source/common/network/dns_resolver/dns_factory_util.h"
 #include "source/common/protobuf/message_validator_impl.h"
 #include "source/common/quic/quic_stat_names.h"
 #include "source/common/router/context_impl.h"
