+++ test/mocks/network/BUILD
@@ -59,7 +59,7 @@ envoy_cc_mock(
         "//source/common/network:address_lib",
         "//source/common/network:socket_interface_lib",
         "//source/common/network:utility_lib",
-        "//source/common/network/dns_resolver:dns_factory_lib",
+        "//source/common/network/dns_resolver:dns_factory_util_lib",
         "//source/common/stats:isolated_store_lib",
         "//test/mocks/event:event_mocks",
         "//test/mocks/stream_info:stream_info_mocks",
